requirements.txt
bcrypt==4.0.1
cffi==1.16.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
fonttools==4.44.0
imageio==2.32.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
paramiko==3.3.1
Pillow==10.0.1
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
scp==0.14.5
scpclient==0.7
six==1.16.0
sklearn==0.0.post11
threadpoolctl==3.2.0
tifffile==2023.9.26
tqdm==4.66.1
tzdata==2023.3
